% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{as_numeric}
\alias{as_numeric}
\alias{as_numeric.default}
\alias{as_numeric.factor}
\title{Converting factors to numeric}
\usage{
as_numeric(x)

\method{as_numeric}{default}(x)

\method{as_numeric}{factor}(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric vector of the same length as x
}
\description{
Converting factors to numeric
}

