% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catf}
\alias{catf}
\title{Print the contents of a file}
\usage{
catf(f)
}
\arguments{
\item{f}{a filename passed to \code{\link{readLines}}.}
}
\value{
The contents of \code{f} are displayed as output.
}
\description{
Wraps \code{\link{cat}} to simple print the contents of a file as output.
}
\examples{
\dontrun{
writeLines(text = c("a", "b", "c"), con = "foo")
catf("foo")
}

}
\references{
Attributed to Yihui Xie via StackOverflow:
\url{http://stackoverflow.com/a/29264573/3277821}
}

