% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{chomp}
\alias{chomp}
\title{Remove trailing newline delimeters from a character vector}
\usage{
chomp(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a version of x without trailing newline delimiters
}
\description{
Remove trailing newline delimeters from a character vector
}
\examples{
chomp("foo\\n")
chomp("bar")

}

