% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear}
\alias{clear}
\title{Remove all objects in the current environment}
\usage{
clear()
}
\description{
An alias for \code{rm(list = ls())}.  Use with care.
}
\examples{
x <- 1
y <- 2
x <- 3
ls()
clear()
ls()

}

