% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{Passed directly to \code{sd} and \code{mean}.}
}
\value{
a scalar equal to sd(x) / mean(x).
}
\description{
Calculate the coefficient of variation (i.e. variation coefficient) for
a vector.
}
\examples{
x <- 1:5
y <- c(x^2, NA)
cv(x) > cv(y, na.rm = TRUE)

}
\references{
Weisstein, Eric W. "Variation Coefficient." From MathWorld--A
Wolfram Web Resource.
\url{http://mathworld.wolfram.com/VariationCoefficient.html}
}

