% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demean.R
\name{group_demean}
\alias{group_demean}
\alias{group_demean.numeric}
\alias{group_demean.formula}
\title{Subtract mean values by grouping factor}
\usage{
group_demean(...)

\method{group_demean}{numeric}(x, groups, digits = getOption("digits"),
  named = FALSE, ...)

\method{group_demean}{formula}(f, data, digits = getOption("digits"),
  named = FALSE, ...)
}
\arguments{
\item{...}{arguments passed to specific methods.}

\item{x}{a numeric vector.}

\item{groups}{a vector of the same length as \code{x} that contains a group indicator 
corresponding to each element of \code{x}.}

\item{digits}{an integer passed directly to \code{\link{round}}.}

\item{named}{If \code{TRUE}, the names of the ouput vector are set to the original 
values of the numeric input data.}

\item{f}{a formula.  For use with the \code{data.frame} \code{data}.  Syntax is 
\code{x ~ groups} where \code{x} is a numeric column of \code{data} that is grouped
by \code{groups}.}

\item{data}{a \code{data.frame}.}
}
\value{
A vector of the same length as the input data (either \code{x} or the input 
column from the \code{data.frame} \code{data}) where group means have been subtracted 
from each element in the input data.  Specifically, residuals are reported the ordinary
least squares regression \code{lm(x ~ factor(groups))}.
}
\description{
Subtract mean values by grouping factor. Accepts either two vectors (data and groups) or 
a \code{data.frame} and a formula of the form \code{data ~ groups}.
}
\examples{
# using a numeric vector and corresponding grouping vector
x <- runif(50)
g <- rep(letters[1:5], 10)
group_demean(x, g)
# using a formula and data.frame
group_demean(mpg ~ cyl, mtcars)

}
