% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haversine.R
\name{haversine_dist}
\alias{haversine_dist}
\title{Calculate distance between geographic coordinate pairs}
\usage{
haversine_dist(l1, l2, r = 6367.4447)
}
\arguments{
\item{l1}{a n x 2 matrix of geographic coordinates.  Each row represents a
point with latitude in the first column and longitude in the second column.}

\item{l2}{a matrix with a schema identical to l1.}

\item{r}{a number for the radius of the reference sphere.  Default is the
average radius of Earth in kilometers
(\url{https://www.wolframalpha.com/input/?i=radius+of+earth})}
}
\value{
a vector of length n where each element is the distance between a
point in l1 and its corresponding point in l2.
}
\description{
Calculate distances between sets of latitude and longitude coordinates using
the law of haversines.  Rough approximation that assumes a spherical Earth.
}
\examples{
l1 <- matrix(runif(20), 20, 2)
l2 <- matrix(runif(20), 20, 2)
haversine_dist(l1, l2)

}
\references{
\url{https://en.wikipedia.org/wiki/Haversine_formula}
}

