% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{is_char_numeric}
\alias{is_char_numeric}
\title{Logical operators}
\usage{
is_char_numeric(x)
}
\arguments{
\item{x}{a vector}
}
\value{
TRUE if \code{x} can be converted into a numeric vector by 
\code{as.numeric} without coercing character elements without a 
commonly-understood numeric representation.
}
\description{
Test if character vector can be converted to numeric without loss of 
information.
}
\examples{
x <- c(1, 3, "3", NA)
xx <- c(x, "A")
is_char_numeric(x)
\dontrun{
is_char_numeric(xx) # FALSE
}
}

