% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{logical}
\alias{logical}
\alias{is_char_numeric}
\alias{is_formula}
\title{Logical operators}
\usage{
is_char_numeric(x)

is_formula(x)
}
\arguments{
\item{x}{an object to check.}
}
\value{
logical
}
\description{
Test an object \code{x}.  See Details for an explanation of how each function operates.
}
\details{
\describe{
    \item{\code{is_char_numeric}}{TRUE if \code{x} can be converted into a numeric vector by \code{as.numeric} without coercing character elements without a commonly-understood numeric representation.}
    \item{\code{is_formula}}{TRUE if \code{x} is a \code{formula} object.}
}
}
\examples{
x <- c(1, 3, "3", NA)
xx <- c(x, "A")
is_char_numeric(x)
f <- "y ~ x"
is_formula(f)
ff <- as.formula(f)
is_formula(ff)
\dontrun{
is_char_numeric(xx) # FALSE
}

}
