% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matches}
\alias{matches}
\title{Return matching elements in a vector}
\usage{
matches(x, pattern = ".", ...)
}
\arguments{
\item{x}{a character vector with elements to be matched}

\item{pattern}{a regular expression.  If \code{length(pattern) > 1}, each element 
is concatenated and a search using \href{https://en.wikipedia.org/wiki/Regular_expression#Basic_concepts}{alternation}.}

\item{...}{additional arguments passed to \code{\link{grepl}}.}
}
\value{
a character vector with the elements of \code{x} that match \code{pattern}
}
\description{
Return matching elements in a vector
}
\examples{
xx <- c("foo", "bar", "baz")
matches(x = xx)
matches(x = xx, pattern = "foo")
matches(x = xx, pattern = "(foo|bar)")
# should be equivalent to
matches(x = xx, pattern = c("foo", "bar"))
matches(x = xx, pattern = "^b.*")

}

