% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfuns.R
\name{most_freq}
\alias{most_freq}
\alias{most_freq.character}
\alias{most_freq.factor}
\alias{most_freq.numeric}
\title{Statistical Mode}
\usage{
most_freq(x, na.rm = TRUE)

\method{most_freq}{numeric}(x, na.rm = TRUE)

\method{most_freq}{character}(x, na.rm = TRUE)

\method{most_freq}{factor}(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{logical If TRUE, all NA values will be dropped.}
}
\value{
a scalar equal to the modal value of \code{x}.
}
\description{
Calculate most frequently occuring value in a vector.
}
\examples{
x <- c(0, 1, 2, 3, 4, 5, 0)
y <- c("foo", "bar", "baz", "foo")
z <- factor(y)
most_freq(x)
most_freq(y)
most_freq(z)
xx <- c(1, x)
most_freq(xx)

}
\references{
\url{http://mathworld.wolfram.com/Mode.html}
}

