% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n}
\alias{n}
\title{Count the number of non-NA observations in a vector}
\usage{
n(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the number of non-NA observations in x as an integer.
}
\description{
Count the number of non-NA observations in a vector
}

