% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ndim}
\alias{ndim}
\title{Count the number of dimensions in an object}
\usage{
ndim(x)
}
\arguments{
\item{x}{a object}
}
\value{
the number dimensions in \code{x}
}
\description{
Count the number of dimensions in an object
}

