% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Flexible vector normalization}
\usage{
normalize(x, fun = c("scale", "unit.length", "unit.range"), ...)
}
\arguments{
\item{x}{a numeric vector}

\item{fun}{either a character vector of length one equal to one of \code{'scale'}, 
\code{'unit.range'} (see Details), or \code{''}, or a function of \code{x} defined 
such that the parameters in \code{...} are passed through.  For example, 
\code{fun = function(x, ...) log(x, ...)}.}

\item{...}{additional arguments passed to \code{fun}.}
}
\value{
a numeric vector of length equal to \code{x} where each original element
of \code{x} has been transformed according to the specification in \code{fun}.
}
\description{
Rescale (normalize) a vector
}
\details{
\describe{
  \item{\code{'scale'}}{\code{base::scale(x, center = TRUE, scale = TRUE)}}
  \item{\code{'unit.range'}}{\code{function(x, ...) = (x - min(x, ...)) / (max(x, ...) - min(x, ...))}}
  \item{\code{'unit.length'}}{\code{function(x) = x / norm(x, type = "2")}}
}
}
\references{
\url{https://en.wikipedia.org/wiki/Normalization_(statistics)}
}
