% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown-templates.R
\name{notes_pdf}
\alias{notes_pdf}
\title{R Markdown Notes Template (PDF)}
\usage{
notes_pdf(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris",
  "+header_attributes", "+auto_identifiers", "+fenced_code_blocks",
  "+fenced_code_attributes", "+fancy_lists", "+subscript", "+superscript",
  "+raw_tex", "+citations", "+inline_notes", "+footnotes", "+table_captions",
  "+pipe_tables"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Attribution: \url{https://github.com/rstudio/rticles}.
}

