% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{pkg}
\alias{in_cran}
\alias{is_installed}
\alias{lib}
\alias{loaded_pkgs}
\alias{pkg}
\title{Package utility functions}
\usage{
loaded_pkgs()

is_installed(pkg)

lib(pkg, ...)

in_cran(pkg)
}
\arguments{
\item{pkg}{a package name either as a string or as a bare (i.e. unquoted) 
name.}

\item{...}{additional arguments passed to either \code{\link{install.packages}}
or \code{\link{library}}.  Properly named arguments will be passed to the
appropriate function.}
}
\description{
Several convenience functions when working with loading, installing, or
querying package database.
}
\details{
\code{loaded_pkgs()} returns a vector of attached packages.
\code{is_installed(pkg)} returns a logical value indicating if \code{pkg} is
installed.  \code{lib(pkg)} checks to see if \code{pkg} is installed,
installs \code{pkg} if it is not, then loads \code{pkg}.  \code{in_cran(pkg)}
returns \code{TRUE} if \code{pkg} is listed on CRAN.
}
\examples{
\dontrun{
loaded_pkgs()
is_installed(ggplot2)
is_installed(stats) == ("stats" \%in\% loaded_pkgs())
lib(downloader)
"downloader" \%in\% loaded_pkgs()
remove.packages("downloader")
in_cran(ggplot2)
in_cran("foo")
}

}

