% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_text}
\alias{read_text}
\title{Data I/O convenience functions}
\usage{
read_text(text, ...)
}
\arguments{
\item{text}{a character vector of length 1.  Should resemble the contents of
a CSV file.}

\item{...}{additional arguments passed directly to \code{read.csv}}
}
\value{
a data.frame
}
\description{
Data I/O convenience functions
}
\examples{
t <- "id,x,y
      a,1,2
      b,3,4"
read_text(t, strip.white = TRUE)
t <- "id,x,y\\na,1,2\\nb,3,4"
read_text(t)

}
