% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a numeric vector}
\usage{
rescale(x, method = "standardize", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{method}{character string specifying the type of scaling.  Currently 
supported options include 'standardize', 'unit.range', and 'unit.length'.}

\item{...}{additional parameters passed to specific rescaling functions 
(e.g. \code{na.rm}).  Note that if \code{x} contains \code{NA}s, 
\code{rescale(x, method = "standardize", na.rm = TRUE)} and \code{rescale(x, 
method = "unit.range")} will not drop \code{NA}s in the resulting scaled 
vector while \code{rescale(x, method = "standardize")} will stop with an
error.}
}
\value{
a numeric vector of equal length as x, rescaled according to the 
procedure chosen in 'type'.
}
\description{
Rescale a numeric vector
}
\details{
Supported rescaling methods:
\describe{
  \item{standardize}{scale(x, ...)}
  \item{unit.range}{(x - min(x, ...)) / (max(x, ...) - min(x, ...))}
  \item{unit.length}{x / norm(x, type = "2")}
}
}
\examples{
test <- c(NA, 1, 2, 3)
rescale(test)
rescale(test, "unit.range", na.rm = TRUE)
test <- test[!is.na(test)]
rescale(test, "unit.length")
\dontrun{ 
# Will fail for undefined rescale method
rescale(test, "foo")
# Will fail on non-numeric vectors
rescale(letters)
# 'unit.length' will fail with NA
rescale(test, "unit.length", na.rm = TRUE)
}
 

}

