% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape_stata}
\alias{long_to_wide}
\alias{reshape_stata}
\alias{wide_to_long}
\title{Reshape the Stata way}
\usage{
wide_to_long(data, stub, i, j, clean = TRUE)

long_to_wide(data, stub, i, j, clean = TRUE)
}
\arguments{
\item{data}{a data.frame}

\item{stub}{a character string.  For \code{wide_to_long}, this is the
leading character "stub" common to all varying columns to be collected into
rows.  For \code{long_to_wide}, this is the name of the variable to be
spread into columns.}

\item{i}{a string}

\item{j}{a string}

\item{clean}{a boolean value.  If TRUE, the resulting data.frame is tidied
automatically: attribute information from \code{reshape} is removed, row
names are reset to a numeric sequence, and column names are stripped of
dots.  Set to FALSE if you would like the immediate result from
\code{reshape}.}
}
\value{
a data.frame.
}
\description{
Wraps the \code{R} function \code{reshape} to mimic the behavior of the 
Stata \code{reshape}.  Used to transform data.frames between "wide" and
"long" formats.
}
\examples{
t <- "id,sex,inc80,inc81,inc82
1,0,5000,5500,6000
2,1,2000,2200,3300
3,0,3000,2000,1000"
df1 <- read.csv(text = t)
df2 <- wide_to_long(data = df1, stub = "inc", i = "id", j = "year")
df3 <- long_to_wide(data = df2, stub = "inc", i = "id", j = "year")
identical(df1, df3)
df4 <- long_to_wide(data = df2, stub = "inc", i = "id", j = "year",
                    clean = FALSE)
identical(df1, df4)

}
\references{
\url{http://www.stata.com/manuals13/dreshape.pdf}
}

