% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_df.R
\name{sample_df}
\alias{sample_df}
\title{Randomly sample a data.frame}
\usage{
sample_df(data, n)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{n}{Either a positive integer or a fraction between 0 and 1}
}
\value{
a subset of randomly selected rows from \code{data}
}
\description{
Sample rows by either specifying the number of desired rows as an integer
(e.g. n = 10) or as a fraction (e.g. n = 0.1) of the total number of rows.
}
\examples{
sample_df(mtcars, n = 10)
sample_df(mtcars, n = 0.1)

}
\seealso{
\code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{sample_frac}}.
}

