% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale-funs.R
\name{scaling-functions}
\alias{scaling-functions}
\alias{center}
\alias{unit_range}
\alias{unit_length}
\title{Numeric vector rescaling functions}
\usage{
center(x, ...)

unit_range(x, ...)

unit_length(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{additional arguments passed to \code{\link{scale}} for \code{center}
or \code{\link{max}} and \code{\link{min}} for \code{unit_range}.}

\item{na.rm}{if \code{TRUE}, \code{unit_length} calculates the L2 norm based on the non-missing
elements of \code{x}.}
}
\value{
a rescaled version of \code{x}.
}
\description{
Rescale a numeric vector
}
\details{
\describe{
  \item{\code{center}}{\eqn{\frac{x - mean(x)}{sd(x)}}}
  \item{\code{unit_range}}{\eqn{\frac{x - min(x)}{max(x) - min(x)}}}
  \item{\code{unit_length}}{\eqn{\frac{x}{\lVert x \rVert}}}
}
}
\examples{
center(1:10)
unit_range(1:10)
unit_length(1:10)

}
\references{
\url{https://en.wikipedia.org/wiki/Normalization_(statistics)}
}
