% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle an object}
\usage{
shuffle(x)
}
\arguments{
\item{x}{Either a vector, list, matrix, or \code{data.frame}}
}
\value{
an object of the same type as \code{x} with elements ordered in a 
random permutation.
}
\description{
Shuffle an object
}
\examples{
shuffle(1:5)
shuffle(letters[1:5])
shuffle(as.list(letters[1:5]))
shuffle(diag(5))
shuffle(mtcars)

}

