% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{strings}
\alias{strings}
\alias{trim}
\alias{chomp}
\alias{text_wrap}
\alias{translit}
\alias{rand_string}
\title{String / text processing functions}
\usage{
trim(x)

chomp(x)

text_wrap(x, n = 20L)

translit(x, ...)

rand_string(n)
}
\arguments{
\item{x}{a character vector}

\item{n}{an integer of length 1.  \code{text_wrap()} will silently coerce
any numeric value passed to \code{n} to an integer.}

\item{...}{addtional parameters used by specific functions.  See Details.}
}
\description{
String / text processing functions
}
\details{
\itemize{
  \item{\code{trim}}{returns a version of \code{x} without trailing OR leading spaces.}
  \item{\code{chomp}}{returns a version of \code{x} without trailing newline delimiters.}
  \item{\code{text_wrap}}{returns a version of x in which each element substitutes newline 
  delimiters for the immediate next space (e.g. ' ') after every \code{n}th character.}
  \item{\code{translit}}{returns a version of \code{x} recoded to ASCII, using \code{iconv} and 
  \code{//TRANSLIT}.  Parameters in \code{...} are passed directly to \code{iconv}.}
  \item{\code{rand_string}}{returns a randomly generated alphanumeric string of length \code{n}.}
}
}
\examples{
trim("foo ")
trim(" bar ")
trim("baz")
chomp("foo\\n")
chomp("bar")
s <- "The quick brown fox jumped over the lazy dog."
text_wrap(s)
text_wrap(s, 10)
cat(paste0(text_wrap(s), "\\n"))
rand_string()
rand_string(5)

}
\references{
I attribute the regular expression that powers \code{text_wrap()} 
to StackOverflow user xiechao: 
\url{http://stackoverflow.com/a/2352006/3277821}
}
