% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{sumstats}
\alias{sumstats}
\title{Simple summary statistics}
\usage{
sumstats(data, f = NULL, digits = 2, order = TRUE)
}
\arguments{
\item{data}{a data.frame}

\item{f}{a formula (See \code{\link[stats]{aggregate}} for details).}

\item{digits}{an integer to specify the number of decimal places desired for
the resulting summary statistics.  Passed directly to \code{round}.}

\item{order}{if TRUE, the rows of the resulting data.frame will be sorted
alphabetically.  If FALSE, the rows will be sorted alphabetically according
to the permutations of the grouping factors in by.}
}
\value{
a data.frame
}
\description{
Generate a simple set of summary statistics for the numeric columns of a
data.frame over various subsets.  There are MANY other ways to do this but
this function is suitable to my needs.
}
\examples{
sumstats(mtcars)
sumstats(iris)
sumstats(iris, cbind(Petal.Width, Sepal.Width) ~ NULL)
sumstats(iris, . ~ Species)
sumstats(mtcars, cbind(mpg, wt) ~ cyl + vs)
sumstats(mtcars, . ~ cyl + vs)
}

