% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symdiff}
\alias{symdiff}
\title{Symmetric set difference}
\usage{
symdiff(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector of the same type as \code{x}.}
}
\value{
the symmetric difference of set \code{x} and \code{y}.
}
\description{
Symmetric set difference
}
\examples{
x <- c(1, 2, 3, 4)
y <- c(1, 4, 5)
symdiff(x, y)
x <- letters[1:4]
y <- letters[2:6]
symdiff(x, y)

}
\references{
Weisstein, Eric W. "Symmetric Difference." 
From MathWorld--A Wolfram Web Resource. 
\url{http://mathworld.wolfram.com/SymmetricDifference.html}
}
