% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_dir}
\alias{temp_dir}
\title{Create a temporary directory}
\usage{
temp_dir(nchar = 10L)
}
\arguments{
\item{nchar}{integer length of random string used as directory name.}
}
\value{
the full directory path of the temporary directory as a string.
}
\description{
Creates a temporary subdirectory of \code{tempdir()}.
}
