% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{text_wrap}
\alias{text_wrap}
\title{Wrap a string using newline delimeters}
\usage{
text_wrap(x, n = 20L)
}
\arguments{
\item{x}{a character vector}

\item{n}{an integer of length 1.  \code{text_wrap()} will silently coerce
any numeric value passed to \code{n} to an integer.}
}
\value{
a version of x in which each element substitutes newline delimiters
for the immediate next space (e.g. ' ') after every \code{n}th character.
}
\description{
Wrap a string using newline delimeters
}
\examples{
s <- "The quick brown fox jumped over the lazy dog."
text_wrap(s)
text_wrap(s, 10)
cat(paste0(text_wrap(s), "\\n"))

}
\references{
I attribute the regular expression that powers \code{text_wrap()} 
to StackOverflow user xiechao: 
\url{http://stackoverflow.com/a/2352006/3277821}
}

