% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{trim}
\alias{trim}
\title{Remove trailing and leading spaces from a character vector}
\usage{
trim(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a version of \code{x} without trailing OR leading spaces
}
\description{
Remove trailing and leading spaces from a character vector
}
\examples{
trim("foo ")
trim(" bar ")
trim("baz")

}
\seealso{
\code{\link[stringr]{str_trim}}
}

