% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear}
\alias{clear}
\title{Remove all objects in an environment}
\usage{
clear(envir = globalenv())
}
\arguments{
\item{envir}{an environment from which all objects will be removed.  Default
is \code{.GlobalEnv} (i.e. ''the user's workspace'').}
}
\description{
An alias for \code{rm(list = ls())} for a specified environment. Use with 
care.
}
\examples{
x <- 1
y <- 2
x <- 3
ls()
clear()
ls()

}
