% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-funs.R
\name{most_freq}
\alias{most_freq}
\alias{most_freq.numeric}
\alias{most_freq.character}
\alias{most_freq.factor}
\title{Calculate most frequently occuring value in a vector}
\usage{
most_freq(x, ...)

\method{most_freq}{numeric}(x, ...)

\method{most_freq}{character}(x, ...)

\method{most_freq}{factor}(x, ...)
}
\arguments{
\item{x}{a vector.}

\item{...}{additional parameters passed on to \code{table}.}
}
\value{
a scalar equal to the modal value of \code{x}.  If 
\code{x} is multimodal, \code{length(x)} will equal the number 
of unique modes in \code{x}.
}
\description{
Calculate most frequently occuring value in a vector (statistical mode).
}
\examples{
x <- c(0, 1, 2, 3, 4, 5, 0)
y <- c("foo", "bar", "baz", "foo")
z <- factor(y)
most_freq(x)
most_freq(y)
most_freq(z)
xx <- c(1, x)
most_freq(xx)

}
\references{
\url{http://mathworld.wolfram.com/Mode.html}
}
