% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a numeric vector}
\usage{
rescale(x, method = "standardize", ...)
}
\arguments{
\item{x}{a numeric vector}

\item{method}{character string specifying the type of scaling.  Currently 
supported options include 'standardize', 'unit.range', and 'unit.length'.}

\item{...}{additional parameters passed to specific rescaling functions 
(e.g. \code{na.rm}).}
}
\value{
a numeric vector of equal length as x, rescaled according to the 
procedure chosen in 'type'.
}
\description{
Rescale a numeric vector
}
\details{
Supported rescaling methods:
\itemize{
  \item 'standardize': scale(x, ...)
  \item 'unit.range': (x - min(x, ...)) / (max(x, ...) - min(x, ...))
  \item 'unit.length' x / sqrt(sum(x^2, ...))
}
}
\examples{
test <- c(NA, 1, 2, 3)
\dontrun{ 
# Will fail for undefined rescale method
rescale(test, "foo")
# Will fail on non-numeric vectors
rescale(letters)
}
rescale(test)
rescale(test, "unit.range", na.rm = TRUE)
rescale(test, "unit.length", na.rm = TRUE)

}

