% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape_stata}
\alias{reshape_stata}
\alias{wide_to_long}
\alias{long_to_wide}
\title{Reshape the Stata way}
\usage{
wide_to_long(data, stub, i, j, clean = TRUE)

long_to_wide(data, stub, i, j, clean = TRUE)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{stub}{a character string.  For \code{wide_to_long()}, this is the
leading character "stub" common to all varying columns to be collected into
rows.  For \code{long_to_wide()}, this is the name of the variable to be
prefixed to the values of \code{j} and spread into columns.}

\item{i}{a string.  The name of the column in \code{data} that uniquely
identifies each observation.}

\item{j}{a string.  For \code{wide_to_long()}, this will be the name for the
new column that captures the values of the varying colums that follow the
value passed to \code{stub}.  For \code{long_to_wide()}, this is the name of
an existing column, the values of which we be suffixed to \code{stub} as the
observations are spread into 'wide' format.}

\item{clean}{a logical value.  If TRUE, the resulting \code{data.frame} is 
tidied automatically: attribute information from \code{reshape()} is 
removed, row names are reset to a numeric sequence, and column names are 
stripped of dots.  Set to FALSE if you would like the immediate result from
\code{reshape()}.}
}
\value{
a \code{data.frame}.
}
\description{
Wraps the function \code{base::reshape()} to mimic the behavior of the 
Stata \code{reshape}.  Used to transform data.frames between "wide" and
"long" formats.
}
\details{
\code{wide_to_long()} and \code{long_to_wide()} are designed to
closely mimic basic functionality of Stata's \code{reshape} command.  See
the examples below for intended usage.
}
\examples{
t <- "id,sex,inc80,inc81,inc82
1,0,5000,5500,6000
2,1,2000,2200,3300
3,0,3000,2000,1000"
df1 <- read.csv(text = t)
df2 <- wide_to_long(data = df1, stub = "inc", i = "id", j = "year")
df3 <- long_to_wide(data = df2, stub = "inc", i = "id", j = "year")
identical(df1, df3)
df4 <- long_to_wide(data = df2, stub = "inc", i = "id", j = "year",
                    clean = FALSE)
identical(df1, df4)

}
\references{
\url{http://www.stata.com/manuals13/dreshape.pdf}
}
