% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle an object}
\usage{
shuffle(x)
}
\arguments{
\item{x}{Either a vector, list, matrix, or data.frame}
}
\value{
an object with the elements of x ordered in a random permutation
}
\description{
Shuffle an object
}
\examples{
shuffle(1:6)
shuffle(letters[1:6])
shuffle(diag(6))
shuffle(mtcars)

}

