% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{sumstats}
\alias{sumstats}
\alias{sumstats.data.frame}
\alias{sumstats.numeric}
\title{Simple summary statistics}
\usage{
sumstats(data, ...)

\method{sumstats}{numeric}(data, digits = 2, ...)

\method{sumstats}{data.frame}(data, f = NULL, digits = 2, order = TRUE,
  ...)
}
\arguments{
\item{data}{a \code{data.frame} or numeric vector.}

\item{...}{additional arguments passed onto various methods.}

\item{digits}{an integer to specify the number of decimal places desired for
the resulting summary statistics.  Passed directly to \code{round()}.}

\item{f}{a formula. See the examples and \code{\link[stats]{aggregate}} for 
details.}

\item{order}{a logical value.  If TRUE (default), the rows of the resulting
\code{data.frame} will be first sorted alphabetically by the variables
specified in the left-hand side of \code{f} then according to the 
permutations of the grouping factors specified in the right-hand side of 
\code{f}.  If FALSE, the rows will be sorted only by permutations of the
grouping factors.}
}
\value{
a \code{data.frame} in which each row represents a variable-grouping
permutation.
}
\description{
Generate a simple set of summary statistics for a numeric vector or the 
numeric columns of a data.frame over various subsets.  There are MANY other 
ways to do this but this function is suitable to my needs.
}
\examples{
# Numeric vector
sumstats(mtcars$disp)
# All variables summarized and no grouping
sumstats(mtcars)
sumstats(iris)
# Only 'Petal.Width' and 'Sepal.Width' summarize and no grouping
sumstats(iris, cbind(Petal.Width, Sepal.Width) ~ NULL)
# All variables summarized and grouped by 'Species'
sumstats(iris, . ~ Species)
# 'mpg' and 'wt' summarized by 'cyl' and 'vs'
sumstats(mtcars, cbind(mpg, wt) ~ cyl + vs)
sumstats(mtcars, cbind(mpg, wt) ~ cyl + vs, order = FALSE)
}

