% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_dir}
\alias{temp_dir}
\title{Create a temporary directory}
\usage{
temp_dir(path = tempdir(), nchar = 10L)
}
\arguments{
\item{path}{base directory in which temporary directory will be created.  Defaults
to \code{\link{tempdir}}.}

\item{nchar}{integer length of random string used as directory name.  Default is
\code{10}.}
}
\value{
the full directory path of the temporary directory as a string.
}
\description{
Creates a temporary subdirectory of \code{\link{tempdir}}.
}
\examples{
tmp <- temp_dir()
tmp
dir.exists(tmp)
unlink(tmp, force = TRUE)

}
